chat/requirements_chat.txt
# Dependências para o Chat RAG
openai>=1.0.0
chromadb>=0.4.0
python-dotenv>=1.0.0
streamlit>=1.28.0
numpy>=1.24.0
pandas>=2.0.0

